implemented wrench2torques() and twist2vels() functions

+++ RaspberryPi/master/kinematics/kinematic.h
@@ -43,7 +43,7 @@ int8_t constraintJacobian(double *Jc, float *qa, float *qu);
 int8_t subchainJacobian(double *Js, float *qa, float *qu, uint8_t chainOption);
 
 // task space to joint space conversions:
-int8_t twist2vels(float *qa, float *qu, float *dqa_dt, float *twist);
-int8_t wrench2torques(float *qa, float *qu, float *torques, float *wrench);
+int8_t twist2vels(float *qa, float *qu, double *dqa_dt, double *twist);
+int8_t wrench2torques(float *qa, float *qu, double *torques, double *wrench);
 
 #endif
